Devided gstliveprofiler and visualizeutil
